refactor: extract a new class from Strings
- The Strings class had functions to manipulate names.
- To facilitate easier code changes, the functions related the
  manipulation of names are made into a new class that deals with just
  name manipulations - ManipulateStringNames.
- This helps to add/change utility methods for both strings and names.
- Helps improve cohesion of both the classes (Strings and
  ManipulateStringNames).
